+++ changelog.md
@@ -2,6 +2,8 @@
 
 - Added the command `/ess convert HuskHomes`, allows converting the database from HuskHomes to zEssentials.
 - Added the command `/ess convert AxVaults`, allows converting the database from AxVaults to zEssentials.
+- Added the command `/mailbox give <player> <item> [<amount>]`, Add an item to a player’s mailbox.
+- Added the command `/mailbox giveall <player> <item> [<amount>]`, Add an item to online player’s mailbox.
 - 
 # 1.0.0.4